+++ ProcessOptimizer/utils/utils.py
@@ -1,10 +1,12 @@
 from copy import deepcopy
 from functools import wraps
+from warnings import warn
 
 import numpy as np
 import matplotlib.pyplot as plt
 from scipy.optimize import OptimizeResult
 from scipy.optimize import minimize as sp_minimize
+from scipy.optimize import LinearConstraint as lin_constraint
 
 from joblib import dump as dump_
 from joblib import load as load_
@@ -16,7 +18,7 @@ __all__ = (
 )
 
 
-def create_result(Xi, yi, space=None, rng=None, specs=None, models=None):
+def create_result(Xi, yi, space=None, rng=None, specs=None, models=None, constraints=None):
     """
     Initialize an `OptimizeResult` object.
 
@@ -39,6 +41,9 @@ def create_result(Xi, yi, space=None, rng=None, specs=None, models=None):
 
     * `models` [list, optional]:
         List of fit surrogate models.
+    
+    * `constraints` [Constraints object, optional]:
+        Constraints that apply to the system.
 
     Returns
     -------
@@ -61,6 +66,7 @@ def create_result(Xi, yi, space=None, rng=None, specs=None, models=None):
         res.space = space
         res.random_state = rng
         res.specs = specs
+        res.constraints = constraints
         return res
     models = np.asarray(models)
     results = []
@@ -79,6 +85,7 @@ def create_result(Xi, yi, space=None, rng=None, specs=None, models=None):
         res.space = space
         res.random_state = rng
         res.specs = specs
+        res.constraints = constraints
         results.append(res)
     return results
 
@@ -266,13 +273,41 @@ def expected_minimum(
 
     xs = [res.x]
     if n_random_starts > 0:
-        xs.extend(res.space.rvs(n_random_starts, random_state=random_state))
+        if getattr(res.constraints, "sum_equals", None):
+           # If we have a SumEquals constraint, create samples that respect it
+           xs = []
+           xs.extend(
+               res.constraints.sumequal_sampling(
+                   n_samples=n_random_starts, random_state=random_state
+                   )
+               )
+        else:
+            if res.constraints:
+                warn.warning("Optimizer has constraints which expected_minimum() does not necessarily respect.")
+            # For all other cases (and constraints) we use random sampling
+            xs.extend(res.space.rvs(n_random_starts, random_state=random_state))
     xs = res.space.transform(xs)
     best_x = None
     best_fun = np.inf
-
+    
+    
+    cons = None
+    # Prepare a linear constraint, if applicable
+    if getattr(res.constraints, "sum_equals", None):
+        A = np.zeros((res.space.n_dims, res.space.n_dims))
+        value = res.constraints.sum_equals[0].value
+        for dim in res.constraints.sum_equals[0].dimensions:
+            A[dim, dim] = 1
+        cons = lin_constraint(A, lb=value, ub=value)
+    
+            
     for x0 in xs:
-        r = sp_minimize(func, x0=x0, bounds=res.space.transformed_bounds)
+        r = sp_minimize(
+            func,
+            x0=x0,
+            bounds=res.space.transformed_bounds,
+            constraints=cons,
+        )
         if r.fun < best_fun:
             best_x = res.space.inverse_transform(np.array(r.x).reshape(1, -1))[0]
             best_fun = r.fun